Feat: 컴포넌트 분리 및 css 수정
